remove unnecessary usage of `implicit Node -> Traversal` (#1658)
We have an `implicit Node -> Traversal` which I now believe was not
a good idea in the first place. This PR cleans up most of it's usage,
which are mostly completely unnecessary calls, like
`val List(sameNodeA) = someNodeA.l`

Unfortunately we can't drop the implicit, I added a comment there.

A follow up PR will tighten some of the return types - this is mostly
a refactoring.