Make Code2CpgFixture more flexible. (#1877)
This enables the reused of Code2CpgFixture also in tests setups where
additional configuration values are need for either the language
frontend or the CPG passes.

We achieve this by basically making the TestCpg a combo object where the
LangaugeFrontend trait is mixed in. By then deriving from TestCpg,
arbitrary additional configuration methods can be added.