jssrc2cpg: Fix Call on Closure Definition Name (#2616)
When a closure is defined directly on an export, there is some kind of currying to enable the export to be a function pointer. This means that the "callee" is a function itself which leads to the "name" property to be the whole function body. This fix checks the callee for being a function and resolves it locally using the `functionNodeToNameAndFullName` mapping.
e.g. of bug:
```
Call(
    id = 46774L,
    argumentIndex = 2,
    argumentName = None,
    code = """(function() {
	var count = 0;
	return function() {
		count++;
		return count;
	};
})()""",
    columnNumber = Some(value = 32),
    dispatchType = "DYNAMIC_DISPATCH",
    dynamicTypeHintFullName = ArraySeq(),
    lineNumber = Some(value = 142),
    methodFullName = """import("/Users/dattaprasadmundada/Documents/tmp/jsrepos/OpenWrap/src_new/util"):function() {
	var count = 0;
	return function() {
		count++;
		return count;
	};
}""",
    name = """function() {
	var count = 0;
	return function() {
		count++;
		return count;
	};
}""",
    order = 2,
    signature = "",
    typeFullName = "ANY"
  )
```