remove traversal, attempt 3 (#2784)
The third attempt at removing Traversal, this time with a detailed migration guide in a new changelog. Please do read the `News.md` file. The thing may be too verbose?

This PR is WIP in the sense that its upstream is not in yet -- I think we first need to review how happy we are with the documentation and the amount of breakage (we might want to change some upstream things to further improve compat here?).