Slicing Documentation: Towards Formalizing the Slices (#2849)
* Simplified data-flow slice
* Modified usages slice to include source code, the one that `joernti` [CodeTIDAL5](https://github.com/joernio/type-inference-models/blob/main/data/model_checkpoints/download_codet5_model.sh) uses
* Increased documentation on `JOERN_SLICE.md`

Note: This will be followed up by the TODO's laid out under `JOERN_SLICE.md`. For now, the usage slices here match `dave/joernti` and this is an effort to set a Joern version with a reproducible artifact following the work on [JoernTI](https://github.com/joernio/type-inference-models). A separate repository will be created to showcase this interaction as to not be something to maintain in Joern directly.