rubysrc2cpg: Added External Dependency Download Support (#2816)
Added Support to resolve the `methodFullName` when external dependency present in ruby file. 

Created one passes:
1. `AstPackagePass`: This pass collect all external dependency files ending with `.rb` extension. Parsing and storing its method info into `packageTable`. 

Added `enableDependencyDownload` flag, If provided, then only gem dependencies will be downloaded. Note, this is troublesome to get working on Windows due to some permissions issue so, for now, this is only available on MacOS and Linux.

Here `packageTable` is like cache for storing the method and require info. Furthermore, we only search for the method name in the module mentioned in the file.