[Performance] Replace `.ast` with `CONTAINS` Traversals where Possible (#3781)
Upon ingesting code bases with large methods that contain many child methods, the `.ast` call becomes extremely expensive on the top of the AST. This work reviews the performance of the CPG generation using the code reported in #3778 to inspect various "hanging" points which were largely related `.ast` calls. These `.ast` calls were replaced by their `_.containsOut` equivalent where appropriate to minimize the traversal depths.

Unfortunately, `XTypeRecovery` is still experimental and relies on `.ast`. Future work is planned to move this to `.statement` instead, and recover select symbols from which shortcut edges can be used. If `XTypeRecovery` hangs, the recommendation is to reduce the iterations to `1` or `0`.

Resolves #3778