[jssrc2cpg] - Graceful handling of error situation for jssrc2cpg (#3911)
While processing `.ejs` files, if the astgen throws any error while
parsing the file. We encoutner this error in Finish method where we read
skiped file contents for the report. In case of `.ejs` files as we
preprocess and convert that into `.js` file. And for some reason this
`.js` file is not present at the expected place, we get file not found
exception. I have handled the situation gracefully in order to avoid
breaking the entire process of generating the CPG.