[jimple2cpg] Add maximum depths for recurse unpack "jars in jars" (#3898)
The [PR 3170](https://github.com/joernio/joern/pull/3170) added an option `--recurse` for `jimple2cpg`, but without max-depths limit.

And also, this make the behavior of `jimple2cpg` odd. If you want to scanning a directory with some jar file you have to specify the `--recurse` option. However, `--recurse` should only control whether to unpack "jars in jars" as described.

This PR correct the behavior, we can now scanning a directory directly without `--recurse` option. If want to enable the `recurse`, set the `--recurse`. And I have add an option `--depths` to control the max-depths of recurse, the default value of max-depth is 1, that's meaning all the jars in a jar file will be unpack but the jars in theses jar will no be unpack again (unless you set the `--depths` to 2).