[ruby] when Statements (Switch Statements) #3926 (#4029)
This draft translates ruby `case` expressions into `if-elif-...-else-end` chains. 

The expression matched against is assigned a temporary variable `<tmp-#>` where # is a number. I was not sure on the convention but I wanted to be sure the variable is always fresh.

A single when can contain a list of expressions to match against, which I translate into an `or-expression` if there is more than one. When matching against an expression `mExpr`, it is turned into a condition with `mExpr.=== <tmp-#>`.

This list of expressions can contain a splat at the end, which we aren't handling yet.

It also remains maybe to special case to a switch ast if all match expressions are literals.

TextSpans for the generated intermediate ast are not sane and will need to be considered more carefully.

---------

Co-authored-by: David Baker Effendi <dave@whirlylabs.com>