cdt: download from official releases rather than jenkins (#4101)
https://download.eclipse.org/tools/cdt/releases/11.4/cdt-11.4.0/plugins/

While we're waiting for 8.4 to get officially released, we need the old jar fiddling hack.
That's unfortunate, but better than having commits in master that won't work in the future (e.g. if we need to bisect anything) because the jars disappeared from their CI (which they do after a week).