[php] Expose frontend command line option for builtins signatures file path (#4103)
The php2cpg frontend uses the SetKnownTypes path to set types for known function signatures, like builtin functions. Currently, it reads the known type signature information from a resource file. This makes it difficult to add new type signatures at runtime, or on a per-project basis.

This PR adds a command line option to the PHP frontend to provide a path to a type signatures file at runtime, so that it can be invoked as:

```
./joern-parse --language php target/ -o target.cpg --frontend-args --known-types-file=known_function_signatures.txt
```